Move axis and lanes into new vue file
